fix(zhipu): 修复 Zhipu API 1210 错误 - temperature 精度问题
  问题根因:
  - StreamRequest.temperature 使用 f32 类型无法精确表示 0.7
  - 序列化时变成 0.699999988079071，导致 Zhipu API 返回 1210 错误
  - 通过 curl 测试确认: temperature=0.699999988079071 → 1210 错误

  修复方案:
  - 将 StreamRequest.temperature 从 Option<f32> 改为 Option<f64>
  - f64 可以精确表示 0.7，避免浮点数精度问题
  - 移除不必要的 temperature 精度修复代码

+++ src-tauri/src/harness/api/types.rs
@@ -12,7 +12,9 @@ pub struct StreamRequest {
     pub messages: Vec<Message>,
     pub max_tokens: u32,
     pub system: Option<String>,
-    pub temperature: Option<f32>,
+    /// 🔥 FIX P0: 使用 f64 避免 Zhipu API 1210 错误
+    /// f32 无法精确表示 0.7，会变成 0.699999988079071
+    pub temperature: Option<f64>,
     pub stream: bool,
     /// 🆕 P2: 工具定义列表
     pub tools: Option<Vec<serde_json::Value>>,
